# [Prim] Optimize composite tanh_triple_grad