<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs 
### Description
<!-- Describe what you’ve done -->

Pcard-75624

Optimize composite OP `tanh_triple_grad` via multiplexing variables to reduce duplicated computation.

| before | after |
|:--: | :--: |
| number prim op calls(17): ![image](https://github.com/PaddlePaddle/Paddle/assets/23737287/9fead16c-8b55-472b-b785-f8841462e36b)| number prim op calls(13):  ![image](https://github.com/PaddlePaddle/Paddle/assets/23737287/87687e77-8e7b-4892-a2d0-d1314fd2f211) |
| Avg. 366 us(over 48 `tanh_triple_grad`) | Avg. 302 us (over 48 `tanh_triple_grad`)|